Cython Bindings / Array View + Tree View (#1349)
* Move TreeArrayView to separate file

* Removing type annotations

* Cython bindings for ArrayView and TreeView

* Typed Arrays
