+++ aim/storage/inmemorytreeview.py
@@ -2,10 +2,10 @@ from copy import deepcopy
 
 from aim.storage.types import CustomObjectBase
 from aim.storage.types import AimObject, AimObjectKey, AimObjectPath
-from aim.storage.arrayview import TreeArrayView
+from aim.storage.treearrayview import TreeArrayView
 from aim.storage.treeview import TreeView
 
-from typing import Iterator, Tuple, Union, List
+from typing import Any, Iterator, Tuple, Union, List
 
 
 class InMemoryTreeView(TreeView):
@@ -155,9 +155,10 @@ class InMemoryTreeView(TreeView):
 
     def array(
         self,
-        path: Union[AimObjectKey, AimObjectPath] = ()
+        path: Union[AimObjectKey, AimObjectPath] = (),
+        dtype: Any = None
     ) -> TreeArrayView:
-        return TreeArrayView(self.subtree(path))
+        return TreeArrayView(self.subtree(path), dtype=dtype)
 
     def first(
         self,
