+++ aim/storage/treeview.py
@@ -1,5 +1,3 @@
-from abc import abstractmethod
-
 from aim.storage.types import AimObject, AimObjectKey, AimObjectPath
 
 from typing import TYPE_CHECKING, Any, Iterator, Tuple, Union, List
@@ -26,7 +24,6 @@ class TreeView:
         # Default to:
         return self.view(path, resolve=False)
 
-    @abstractmethod
     def view(
         self,
         path: Union[AimObjectKey, AimObjectPath],
@@ -34,14 +31,12 @@ class TreeView:
     ):
         ...
 
-    @abstractmethod
     def make_array(
         self,
         path: Union[AimObjectKey, AimObjectPath] = ()
     ):
         ...
 
-    @abstractmethod
     def collect(
         self,
         path: Union[AimObjectKey, AimObjectPath] = (),
@@ -66,14 +61,12 @@ class TreeView:
         except KeyError:
             return default
 
-    @abstractmethod
     def __delitem__(
         self,
         path: Union[AimObjectKey, AimObjectPath]
     ):
         ...
 
-    @abstractmethod
     def set(
         self,
         path: Union[AimObjectKey, AimObjectPath],
@@ -82,7 +75,6 @@ class TreeView:
     ):
         self.__setitem__(path, value)
 
-    @abstractmethod
     def __setitem__(
         self,
         path: Union[AimObjectKey, AimObjectPath],
@@ -91,14 +83,12 @@ class TreeView:
         self.set(path, value, strict=True)
 
     # TODO [MV]: revisit this approach after cython bindings refactoring
-    @abstractmethod
     def keys_eager(
             self,
             path: Union[AimObjectKey, AimObjectPath] = (),
     ) -> List[Union[AimObjectPath, AimObjectKey]]:
         ...
 
-    @abstractmethod
     def keys(
         self,
         path: Union[AimObjectKey, AimObjectPath] = (),
@@ -106,7 +96,6 @@ class TreeView:
     ) -> Iterator[Union[AimObjectPath, AimObjectKey]]:
         ...
 
-    @abstractmethod
     def items_eager(
             self,
             path: Union[AimObjectKey, AimObjectPath] = ()
@@ -116,7 +105,6 @@ class TreeView:
     ]]:
         ...
 
-    @abstractmethod
     def items(
         self,
         path: Union[AimObjectKey, AimObjectPath] = ()
@@ -126,32 +114,19 @@ class TreeView:
     ]]:
         ...
 
-    @abstractmethod
-    def iterlevel(
-        self,
-        path: Union[AimObjectKey, AimObjectPath] = (),
-        level: int = 1
-    ) -> Iterator[Tuple[
-        AimObjectPath,
-        AimObject
-    ]]:
-        ...
-
-    @abstractmethod
     def array(
         self,
-        path: Union[AimObjectKey, AimObjectPath] = ()
+        path: Union[AimObjectKey, AimObjectPath] = (),
+        dtype: Any = None
     ) -> 'ArrayView':
         ...
 
-    @abstractmethod
     def first(
         self,
         path: Union[AimObjectKey, AimObjectPath] = ()
     ) -> Tuple[AimObjectKey, AimObject]:
         ...
 
-    @abstractmethod
     def last(
         self,
         path: Union[AimObjectKey, AimObjectPath] = ()
