+++ aim/storage/treeviewproxy.py
@@ -1,11 +1,13 @@
 from aim.ext.transport.message_utils import ResourceObject, pack_args
 from aim.ext.transport.remote_resource import RemoteResourceAutoClean
-from aim.storage.arrayview import TreeArrayView
-from aim.storage.treeutils import encode_tree
+
 from aim.storage.treeview import TreeView
+from aim.storage.treeutils import encode_tree
+from aim.storage.treearrayview import TreeArrayView
 from aim.storage.types import AimObject, AimObjectKey, AimObjectPath
 
-from typing import TYPE_CHECKING, Iterator, Tuple, Union, List
+from typing import TYPE_CHECKING, Any, Iterator, Tuple, Union, List
+
 
 if TYPE_CHECKING:
     from aim.ext.transport.client import Client
@@ -128,9 +130,10 @@ class ProxyTree(TreeView):
 
     def array(
         self,
-        path: Union[AimObjectKey, AimObjectPath] = ()
+        path: Union[AimObjectKey, AimObjectPath] = (),
+        dtype: Any = None
     ) -> TreeArrayView:
-        return TreeArrayView(self.subtree(path))
+        return TreeArrayView(self.subtree(path), dtype=dtype)
 
     def first(
         self,
@@ -233,9 +236,10 @@ class SubtreeView(TreeView):
 
     def array(
         self,
-        path: Union[AimObjectKey, AimObjectPath] = ()
+        path: Union[AimObjectKey, AimObjectPath] = (),
+        dtype: Any = None
     ) -> TreeArrayView:
-        return TreeArrayView(self.subtree(path))
+        return TreeArrayView(self.subtree(path), dtype=dtype)
 
     def first(
         self,
